feat: move _forward_train in dreamzero patch
